bug fixed

+++ src/main/java/com/subhashCart/services/impl/SellerServiceImpl.java
@@ -1,8 +1,14 @@
 package com.subhashCart.services.impl;
 
+import com.subhashCart.dtos.SellerDTO;
+import com.subhashCart.dtos.SessionDTO;
+import com.subhashCart.exceptions.LoginException;
 import com.subhashCart.exceptions.SellerException;
 import com.subhashCart.models.Seller;
+import com.subhashCart.models.UserSession;
 import com.subhashCart.repositories.SellerDao;
+import com.subhashCart.repositories.SessionDao;
+import com.subhashCart.services.LoginLogoutService;
 import com.subhashCart.services.SellerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,6 +21,13 @@ public class SellerServiceImpl implements SellerService {
 
     @Autowired
     private SellerDao sellerDao;
+
+    @Autowired
+    private LoginLogoutService loginService;
+
+    @Autowired
+    private SessionDao sessionDao;
+
 
     @Override
     public Seller addSeller(Seller seller) {
@@ -48,7 +61,13 @@ public class SellerServiceImpl implements SellerService {
     }
 
     @Override
-    public Seller updateSeller(Seller seller) {
+    public Seller updateSeller(Seller seller, String token) {
+
+        if(token.contains("seller") == false) {
+            throw new LoginException("Invalid session token for seller");
+        }
+
+        loginService.checkTokenStatus(token);
 
         Seller existingSeller=sellerDao.findById(seller.getSellerId()).orElseThrow(()-> new SellerException("Seller not found for this Id: "+seller.getSellerId()));
         Seller newSeller= sellerDao.save(seller);
@@ -56,28 +75,135 @@ public class SellerServiceImpl implements SellerService {
     }
 
     @Override
-    public Seller deleteSellerById(Integer sellerId) {
+    public Seller deleteSellerById(Integer sellerId, String token) {
+
+        if(token.contains("seller") == false) {
+            throw new LoginException("Invalid session token for seller");
+        }
+
+        loginService.checkTokenStatus(token);
+
         Optional<Seller> opt=sellerDao.findById(sellerId);
 
         if(opt.isPresent()) {
+
+            UserSession user = sessionDao.findByToken(token).get();
+
             Seller existingseller=opt.get();
-            sellerDao.delete(existingseller);
-            return existingseller;
+
+            if(user.getUserId() == existingseller.getSellerId()) {
+                sellerDao.delete(existingseller);
+
+                // logic to log out a seller after he deletes his account
+                SessionDTO session = new SessionDTO();
+                session.setToken(token);
+                loginService.logoutSeller(session);
+
+                return existingseller;
+            }
+            else {
+                throw new SellerException("Verification Error in deleting seller account");
+            }
         }
         else throw new SellerException("Seller not found for this ID: "+sellerId);
 
     }
 
     @Override
-    public Seller updateSellerMobile(Integer sellerId, String mobile) throws SellerException {
+    public Seller updateSellerMobile(SellerDTO sellerdto, String token) throws SellerException {
 
-        Seller existingSeller=sellerDao.findById(sellerId).orElseThrow(()->new SellerException("Seller not found for this ID: "+sellerId));
+        if(token.contains("seller") == false) {
+            throw new LoginException("Invalid session token for seller");
+        }
 
-        existingSeller.setMobile(mobile);
+        loginService.checkTokenStatus(token);
 
-        return sellerDao.save(existingSeller);
+        UserSession user = sessionDao.findByToken(token).get();
+
+        Seller existingSeller=sellerDao.findById(user.getUserId()).orElseThrow(()->new SellerException("Seller not found for this ID: "+ user.getUserId()));
+
+        if(existingSeller.getPassword().equals(sellerdto.getPassword())) {
+            existingSeller.setMobile(sellerdto.getMobile());
+            return sellerDao.save(existingSeller);
+        }
+        else {
+            throw new SellerException("Error occured in updating mobile. Please enter correct password");
+        }
+
+    }
+
+    @Override
+    public Seller getSellerByMobile(String mobile, String token) throws SellerException {
+
+        if(token.contains("seller") == false) {
+            throw new LoginException("Invalid session token for seller");
+        }
+
+        loginService.checkTokenStatus(token);
+
+        Seller existingSeller = sellerDao.findByMobile(mobile).orElseThrow( () -> new SellerException("Seller not found with given mobile"));
+
+        return existingSeller;
+    }
+
+    @Override
+    public Seller getCurrentlyLoggedInSeller(String token) throws SellerException{
+
+        if(token.contains("seller") == false) {
+            throw new LoginException("Invalid session token for seller");
+        }
+
+        loginService.checkTokenStatus(token);
+
+        UserSession user = sessionDao.findByToken(token).get();
+
+        Seller existingSeller=sellerDao.findById(user.getUserId()).orElseThrow(()->new SellerException("Seller not found for this ID"));
+
+        return existingSeller;
+
     }
 
 
+    // Method to update password - based on current token
+
+    @Override
+    public SessionDTO updateSellerPassword(SellerDTO sellerDTO, String token) {
+
+        if(token.contains("seller") == false) {
+            throw new LoginException("Invalid session token for seller");
+        }
+
+
+        loginService.checkTokenStatus(token);
+
+        UserSession user = sessionDao.findByToken(token).get();
+
+        Optional<Seller> opt = sellerDao.findById(user.getUserId());
+
+        if(opt.isEmpty())
+            throw new SellerException("Seller does not exist");
+
+        Seller existingSeller = opt.get();
+
+
+        if(sellerDTO.getMobile().equals(existingSeller.getMobile()) == false) {
+            throw new SellerException("Verification error. Mobile number does not match");
+        }
+
+        existingSeller.setPassword(sellerDTO.getPassword());
+
+        sellerDao.save(existingSeller);
+
+        SessionDTO session = new SessionDTO();
+
+        session.setToken(token);
+
+        loginService.logoutSeller(session);
+
+        session.setMessage("Updated password and logged out. Login again with new password");
+
+        return session;
+
+    }
 
 }
